requirements.txt
Flask 
google-cloud-bigquery
http
requests
json